[FEAT] 지도 검색 후 메뉴 상세정보 화면으로의 이동 구현

+++ app/src/main/java/com/kuit/ourmenu/data/model/map/response/MapSearchHistoryResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class MapSearchHistoryResponse(
     @SerialName("menuId")
-    val menuId: Int,
+    val menuId: Long,
     @SerialName("menuTitle")
     val menuTitle: String,
     @SerialName("storeTitle")
