+++ app/src/main/java/com/kuit/ourmenu/ui/menuFolder/screen/MenuFolderAllMenuScreen.kt
@@ -56,7 +56,7 @@ import kotlinx.coroutines.launch
 @Composable
 fun MenuFolderAllMenuScreen(
     onNavigateBack: () -> Unit,
-    onNavigateToMenuInfo: (Int) -> Unit,
+    onNavigateToMenuInfo: (Long) -> Unit,
 //    onNavigateToMap: () -> Unit, // TODO: Map으로 화면 이동 구현
     onNavigateToAddMenu: () -> Unit,
     viewModel: MenuFolderAllViewModel = hiltViewModel()
